Refactor to improve toml setup

+++ packages/cli/src/commands/setup/graphql/features/trusted-documents.ts
@@ -3,19 +3,15 @@ import path from 'path'
 
 import execa from 'execa'
 import { Listr } from 'listr2'
-import { format } from 'prettier'
 import { Project, SyntaxKind, type PropertyAssignment } from 'ts-morph'
 
-import {
-  recordTelemetryAttributes,
-  prettierOptions,
-} from '@redwoodjs/cli-helpers'
+import { recordTelemetryAttributes, prettifyFile } from '@redwoodjs/cli-helpers'
 import { getConfigPath } from '@redwoodjs/project-config'
 
 import { getPaths } from '../../../../lib'
 import c from '../../../../lib/colors'
 
-export const command = 'trusted-docs'
+export const command = 'trusted-documents'
 export const description = 'Set up Trusted Documents for GraphQL'
 
 export function builder(yargs: any) {
@@ -43,7 +39,7 @@ function configureGraphQLHandlerWithStore() {
 
       if (!graphQlSourcePath) {
         console.warn(
-          c.warning(
+          c.yellow(
             `Unable to find the GraphQL Handler source file: ${path.join(
               functionsDir,
               'graphql.(js|ts)'
@@ -56,8 +52,20 @@ function configureGraphQLHandlerWithStore() {
       // add import
       const project = new Project()
       const graphQlSourceFile =
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        project.addSourceFileAtPathIfExists(graphQlSourcePath)!
+        project.addSourceFileAtPathIfExists(graphQlSourcePath)
+
+      if (!graphQlSourceFile) {
+        console.error(
+          c.error(
+            `Unable to determine the GraphQL Handler source path for: ${path.join(
+              functionsDir,
+              'graphql.(js|ts)'
+            )}`
+          )
+        )
+        return
+      }
+
       let graphQlSourceFileChanged = false
       let identified = false
 
@@ -118,12 +126,7 @@ function configureGraphQLHandlerWithStore() {
 
       if (graphQlSourceFileChanged) {
         await project.save()
-        const updatedHandler = fs.readFileSync(graphQlSourcePath, 'utf-8')
-        const prettifiedHandler = format(updatedHandler, {
-          ...prettierOptions(),
-          parser: 'babel-ts',
-        })
-        fs.writeFileSync(graphQlSourcePath, prettifiedHandler, 'utf-8')
+        prettifyFile(graphQlSourcePath, 'babel-ts')
       }
     },
   }
@@ -137,7 +140,7 @@ export async function handler({
   install: boolean
 }) {
   recordTelemetryAttributes({
-    command: 'setup graphql trusted-docs',
+    command: 'setup graphql trusted-documents',
     force,
     install,
   })
@@ -151,12 +154,55 @@ export async function handler({
         task: () => {
           const redwoodTomlPath = getConfigPath()
           const originalTomlContent = fs.readFileSync(redwoodTomlPath, 'utf-8')
+          let newConfig = undefined
 
-          const tomlToAppend = `[graphql]\n  trustedDocuments = true`
+          const graphqlExists = originalTomlContent.includes('[graphql]')
 
-          const newConfig = originalTomlContent + '\n' + tomlToAppend
+          const trustedDocumentsExists =
+            originalTomlContent.includes('trustedDocuments =') ||
+            originalTomlContent.includes('trustedDocuments=')
 
-          fs.writeFileSync(redwoodTomlPath, newConfig, 'utf-8')
+          const fragmentsExists =
+            originalTomlContent.includes('fragments =') ||
+            originalTomlContent.includes('fragments=')
+
+          if (trustedDocumentsExists) {
+            console.info(
+              c.white(
+                'GraphQL Trusted Documents are already enabled in your Redwood project.'
+              )
+            )
+          } else if (graphqlExists && fragmentsExists) {
+            const insertIndex = originalTomlContent.indexOf('fragments')
+            const trustedDocuments = 'trustedDocuments = true\n  '
+            newConfig =
+              originalTomlContent.slice(0, insertIndex) +
+              trustedDocuments +
+              originalTomlContent.slice(insertIndex)
+          } else {
+            if (!graphqlExists) {
+              const tomlToAppend = `[graphql]\n  trustedDocuments = true`
+
+              newConfig = originalTomlContent + '\n' + tomlToAppend
+            } else {
+              const graphqlIndex = originalTomlContent.indexOf('[graphql]')
+              const insertIndex = graphqlIndex + '[graphql]\n'.length
+              newConfig =
+                originalTomlContent.slice(0, insertIndex) +
+                '  trustedDocuments = true\n' +
+                originalTomlContent.slice(insertIndex)
+            }
+          }
+
+          if (newConfig && (force || !trustedDocumentsExists)) {
+            fs.writeFileSync(redwoodTomlPath, newConfig, 'utf-8')
+          } else {
+            console.warn(
+              c.yellow(
+                'Unable to update Redwood Project Configuration to enable GraphQL Trusted Documents. Please add it manually following https://docs.redwoodjs.com/docs/graphql/trusted-documents#configure-redwood-project-configuration'
+              )
+            )
+          }
         },
       },
       {
